Agregados chequeos en los servicios -Se agrego la excepcion id not found, para que cuando cuando create o update no encuentren el producto con la id llame (error 404) -Se agregaron errores de bad reques (400) para el caso donde en update se quiera cambiar el id, y para cuando en el create no se rellene alguno de los campos (a excepción el id)

+++ src/app.service.ts
@@ -1,7 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Producto } from './Entity/Producto';
+import { IdNotFound } from './Customs exceptions/id.not.found';
 
 @Injectable()
 export class AppService {
@@ -18,6 +19,9 @@ export class AppService {
 
   async createProducto(p: Producto)
   {
+    if(p.nombre == null || p.precio == null || p.descripcion == null)
+      throw new HttpException('Para añadir un producto los campos de nombre, precio y descripcion no pueden estar vacíos.',
+      HttpStatus.BAD_REQUEST);
     const newP = await this.productosRepository.create(p)
     return this.productosRepository.save(newP);
   }
@@ -25,12 +29,18 @@ export class AppService {
   async updateProducto(id: number, p: Producto)
   {
     const oldP = await this.productosRepository.findOneBy({id: id});
+    if(oldP == null)
+      throw new IdNotFound(id);
+    if(p.id != null)
+      throw new HttpException('No se puede actualizar el id de un producto.', HttpStatus.BAD_REQUEST);
     this.productosRepository.merge(oldP, p)
     return this.productosRepository.save(oldP);
   }
 
   async deleteProducto(id: number)
   {
+    if(await this.productosRepository.findOneBy({id: id}) == null)
+      throw new IdNotFound(id);
     await this.productosRepository.delete(id);
     return true;
   }
